Added exception reporting with Bugout
There are 2 new config parameters:
- `BUGOUT_JOURNAL_ID` - this is the ID of a demo knowledge base I
created to show how error reporting works with Bugout.
- `BUGOUT_ACCESS_TOKEN` - this is an access token for a user who only
has permissions to create new entries in the Activeloop demo knowledge
base.

For the exceptions inheriting from `HubException`, I tried to implement
error reporting in the `HubException.__init__`, but it didn't work
cleanly. I suspect this is because `ClickException` has some magic in it
that I don't understand.